# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CryptoVault Exchange & Shop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List user withdrawals
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Withdrawal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
        "401":
          description: Unauthorized
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Request a USDT withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Withdrawal:
      type: object
      required: [id, userId, amount, fee, netAmount, destinationAddress, status, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        amount: { type: string }
        fee: { type: string }
        netAmount: { type: string }
        destinationAddress: { type: string }
        status:
          type: string
          enum: [pending, approved, rejected]
        txHash: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        processedAt: { type: ["string", "null"] }
    WithdrawalRequest:
      type: object
      required: [amount, destinationAddress]
      properties:
        amount: { type: string }
        destinationAddress: { type: string }
...